Rename module methodscript -> _methodscript
