+++ python/tests/methodscript/test_parser.py
@@ -5,7 +5,7 @@ from antlr4 import CommonTokenStream, InputStream
 from antlr4.error.ErrorListener import ErrorListener
 from typing_extensions import override
 
-from pypalmsens.methodscript import (
+from pypalmsens._methodscript import (
     MethodScriptLexer,
     MethodScriptParser,
 )
